PPTP-828 Update EIS Subscription model
The intention is to use the EIS Subscription model as much as possible,
so that the views render the data according to the EIS model. This saves
us from having to map back and forth between PPT and EIS models.

It becomes more tightly coupled design, but the EIS schema is now been
baselined and it should not change that often.