PPTP-838 Replace CSS #id .class to style rows
**Context:**

Some rows require some border lines in order to provide emphasis on the
values of those same rows
In the past we have used the CSS id `#row-border`, this is a problem
because CSS ids have to be unique per page.

This change will transform the `row-order` into a CSS class.