PPTP-1867 : Delete Successfully Submitted Tax Returns (#154)
We only have a regular AuthenticatedRequest when displaying the return submission confirmation page (as the temporary TaxReturn is now deleted). Since we have commented out sections which appear to need details from the submitted return we might need to make a call to pick up the submitted return from EIS/ETMP before displaying this page and passing this TaxReturn as a parameter to the page - we can fix this if and when it is needed.